backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.39.0
yfinance==0.2.44
chromadb==0.5.23
sentence-transformers==3.0.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
cachetools==5.5.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
redis[hiredis]==5.2.0
pymupdf>=1.25.0
python-docx==1.1.0
slowapi==0.1.9
charset-normalizer==3.3.2

